requirements.txt
pytest
selenium
webdriver-manager
allure-pytest